requirements.txt
beautifulsoup4==4.5.1
boto==2.42.0
dj-static==0.0.6
Django==1.10
django-appconf==1.0.2
django-ckeditor==5.1.0
django-cleanup==0.4.2
django-compressor==2.1
django-extra-views==0.8.0
django-filter==0.14.0
django-gunicorn==0.1.1
django-htmlmin==0.9.1
django-imagekit==3.3
django-storages==1.5.0
django-wysiwyg==0.8.0
djangorestframework==3.4.6
gunicorn==19.6.0
html5lib==0.999999999
pilkit==1.1.13
Pillow==3.3.1
psycopg2==2.6.2
python-dateutil==2.5.3
pyuploadcare==2.1.0
rcssmin==1.0.6
requests==2.11.1
rjsmin==1.0.12
six==1.10.0
static3==0.7.0
webencodings==0.5
whitenoise==3.2.2
